% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dec2019.R
\docType{data}
\name{dec2019}
\alias{dec2019}
\title{UK Ward Boundaries December 2019}
\format{
Simple feature collection with 8887 features and 13 fields.
\describe{
\item{OBJECTID}{Equivalent to row number}
\item{WD19CD}{2019 Ward Code}
\item{WD19NM}{2019 Ward Name (English)}
\item{WD19NMW}{2019 Ward Name (Welsh)}
\item{BNG_E}{Easting}
\item{BNG_N}{Northing}
\item{LONG}{Longitude}
\item{LAT}{Latitude}
\item{Shape__Area}{Surface area (square meters)}
\item{Shape__Length}{Boundary length (meters)}
\item{GlobalID}{Not sure what this one is...}
\item{geometry}{Geometry type of MULTIPOLYGON}
\item{LAD19CD}{2019 Local Authority District Code}
\item{LAD19NM}{2019 Local Authority District Name}
}
}
\source{
\url{https://services1.arcgis.com/ESMARspQHYMw9BZ9/arcgis/rest/services/Wards_December_2019_Boundaries_UK_BGC_2022/FeatureServer/0/query?outFields=*&where=1\%3D1&f=geojson}
}
\usage{
dec2019
}
\description{
Boundaries for Wards in the United Kingdom as at December 2019 The boundaries are: (BGC) Generalised (20m) - clipped to the coastline (Mean High Water mark). LAD codes and names have been added using the ONS Dec 2019 lookup. Contains both Ordnance Survey and ONS Intellectual Property Rights.
}
\keyword{datasets}
