% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lad-may2023-20m.R
\docType{data}
\name{lad_may2023_20m}
\alias{lad_may2023_20m}
\title{UK Local Authority District Boundaries May 2023}
\format{
Simple feature collection with 361 features and 10 fields.
\describe{
\item{FID}{Feature ID}
\item{LAD23CD}{2023 Local Authority District Code}
\item{LAD23NM}{2023 Local Authority District Name}
\item{LAD23NMW}{2023 Local Authority District Name in Welsh}
\item{BNG_E}{Easting}
\item{BNG_N}{Northing}
\item{LONG}{Longitude}
\item{LAT}{Latitude}
\item{Shape__Area}{Surface area (square meters)}
\item{Shape__Length}{Boundary length (meters)}
\item{geometry}{Geometry type of MULTIPOLYGON}
}
}
\source{
\url{https://services1.arcgis.com/ESMARspQHYMw9BZ9/arcgis/rest/services/Local_Authority_Districts_May_2023_UK_BGC/FeatureServer/0/query?outFields=*&where=1\%3D1&f=geojson}
}
\usage{
lad_may2023_20m
}
\description{
Boundaries for Local Authority Districts in the United Kingdom as at May 2023. The boundaries available are: (BGC) Generalised (20m) - clipped to the coastline (Mean High Water Mark). Contains both Ordnance Survey and ONS Intellectual Property Rights.
}
\keyword{datasets}
