% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wards-dec2020.R
\docType{data}
\name{wards_dec2020}
\alias{wards_dec2020}
\title{UK Ward Boundaries December 2020}
\format{
Simple feature collection with 8887 features and 12 fields.
\describe{
\item{OBJECTID}{Equivalent to row number}
\item{WD20CD}{2020 Ward Code}
\item{WD20NM}{2020 Ward Name (English)}
\item{WD20NMW}{2020 Ward Name (Welsh)}
\item{BNG_E}{Easting}
\item{BNG_N}{Northing}
\item{LONG}{Longitude}
\item{LAT}{Latitude}
\item{Shape__Area}{Surface area (square meters)}
\item{Shape__Length}{Boundary length (meters)}
\item{geometry}{Geometry type of MULTIPOLYGON}
\item{LAD20CD}{2020 Local Authority District Code}
\item{LAD20NM}{2020 Local Authority District Name}
}
}
\source{
\url{https://services1.arcgis.com/ESMARspQHYMw9BZ9/arcgis/rest/services/Wards_Dec_2020_UK_BSC_V2_2022/FeatureServer/0/query?outFields=*&where=1\%3D1&f=geojson}
}
\usage{
wards_dec2020
}
\description{
Boundaries for Wards in the United Kingdom as at December 2020. The boundaries are: (BSC) Super Generalised (200m) - clipped to the coastline (Mean High Water mark). LAD codes and names have been added using the ONS Dec 2020 lookup. Contains both Ordnance Survey and ONS Intellectual Property Rights.
}
\keyword{datasets}
