% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wardboundaries-package.R
\docType{package}
\name{wardboundaries-package}
\alias{wardboundaries}
\alias{wardboundaries-package}
\title{wardboundaries: UK Ward Boundaries for Mapping}
\description{
Contains yearly datasets of UK ward boundaries from the ONS Open Geography Portal. Datasets are pre-formatted as sf objects with a geom column of type MULTIPOLYGON.
}
\author{
\strong{Maintainer}: Jim Gardner \email{jimmahgee@gmail.com}

}
\keyword{internal}
